fix bug in PCH algorihtm when its calculates the communication costs

+++ sources/org/workflowsim/DatacenterExtended.java
@@ -334,7 +334,7 @@ public class DatacenterExtended extends Datacenter {
                             }
                         }
                         if (requiredFileStagein && maxBwth > 0.0) {
-                            time += file.getSize() / Consts.MILLION * 8 / maxBwth;
+                            time += file.getSize() / (double)Consts.MILLION * 8 / maxBwth;
                         }
 
                         /**
